Move decoder to playback crate
