+++ .github/workflows/test.yml
@@ -99,7 +99,6 @@ jobs:
       - run: cargo hack  --workspace --remove-dev-deps
       - run: cargo build -p librespot-core --no-default-features
       - run: cargo build -p librespot-core
-      - run: cargo hack build --each-feature -p librespot-audio
       - run: cargo build -p librespot-connect
       - run: cargo build -p librespot-connect --no-default-features --features with-dns-sd
       - run: cargo hack build --each-feature
