+++ playback/Cargo.toml
@@ -18,12 +18,14 @@ path = "../metadata"
 version = "0.2.0"
 
 [dependencies]
+cfg-if = "1.0"
 futures-executor = "0.3"
 futures-util = { version = "0.3", default_features = false, features = ["alloc"] }
 log = "0.4"
 byteorder = "1.4"
 shell-words = "1.0.0"
 tokio = { version = "1", features = ["sync"] }
+zerocopy        = { version = "0.3" }
 
 alsa            = { version = "0.5", optional = true }
 portaudio-rs    = { version = "0.3", optional = true }
@@ -34,12 +36,17 @@ sdl2            = { version = "0.34.3", optional = true }
 gstreamer       = { version = "0.16", optional = true }
 gstreamer-app   = { version = "0.16", optional = true }
 glib            = { version = "0.10", optional = true }
-zerocopy        = { version = "0.3" }
 
 # Rodio dependencies
 rodio           = { version = "0.13", optional = true, default-features = false }
 cpal            = { version = "0.13", optional = true }
 thiserror       = { version = "1", optional = true }
+
+# Decoders
+lewton = "0.10"  # Currently not optional because of limitations of cargo features
+librespot-tremor = { version = "0.2", optional = true }
+ogg = "0.8"
+vorbis = { version ="0.0", optional = true }
 
 [features]
 alsa-backend = ["alsa"]
@@ -50,3 +57,6 @@ rodio-backend = ["rodio", "cpal", "thiserror"]
 rodiojack-backend = ["rodio", "cpal/jack", "thiserror"]
 sdl-backend = ["sdl2"]
 gstreamer-backend = ["gstreamer", "gstreamer-app", "glib"]
+
+with-tremor = ["librespot-tremor"]
+with-vorbis = ["vorbis"]
