+++ playback/src/audio_backend/mod.rs
@@ -1,5 +1,5 @@
-use crate::audio::AudioPacket;
 use crate::config::AudioFormat;
+use crate::decoder::AudioPacket;
 use std::io;
 
 pub trait Open {
@@ -26,7 +26,7 @@ fn mk_sink<S: Sink + Open + 'static>(device: Option<String>, format: AudioFormat
 macro_rules! sink_as_bytes {
     () => {
         fn write(&mut self, packet: &AudioPacket) -> io::Result<()> {
-            use crate::audio::convert::{self, i24};
+            use crate::convert::{self, i24};
             use zerocopy::AsBytes;
             match packet {
                 AudioPacket::Samples(samples) => match self.format {
