+++ playback/src/audio_backend/pulseaudio.rs
@@ -1,6 +1,6 @@
 use super::{Open, Sink, SinkAsBytes};
-use crate::audio::AudioPacket;
 use crate::config::AudioFormat;
+use crate::decoder::AudioPacket;
 use crate::player::{NUM_CHANNELS, SAMPLE_RATE};
 use libpulse_binding::{self as pulse, stream::Direction};
 use libpulse_simple_binding::Simple;
