+++ playback/src/audio_backend/subprocess.rs
@@ -1,6 +1,6 @@
 use super::{Open, Sink, SinkAsBytes};
-use crate::audio::AudioPacket;
 use crate::config::AudioFormat;
+use crate::decoder::AudioPacket;
 use shell_words::split;
 
 use std::io::{self, Write};
